refactor: move WaveRayPath to its own file
